파일 구조 체계화
